# asklepios-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Asklepios ERP API Documentation",
        "description": "Documentation interactive de l'API Asklepios pour la gestion hospitalière. Permet aux développeurs front-end de tester les requêtes en temps réel.",
        "contact": {
            "name": "Support Technique Asklepios",
            "email": "admin@asklepios.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API de Développement"
        }
    ],
    "paths": {
...
        "/api/purchase-orders": {
            "get": {
                "tags": [
                    "Commandes Fournisseurs"
                ],
                "summary": "Lister les commandes (Filtrable avec Pagination)",
                "operationId": "75ef4b95a8b16eeab89a5ded5dbdc9bf",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de la page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrer par statut (ex: PENDING, RECEIVED)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "provider_id",
                        "in": "query",
                        "description": "Filtrer par fournisseur",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des commandes récupérée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Commandes Fournisseurs"
                ],
                "summary": "Créer une commande",
                "operationId": "77591da94b6c5cb5785428116df32962",
                "responses": {
                    "201": {
                        "description": "Commande créée avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/purchase-orders/{id}": {
            "put": {
                "tags": [
                    "Commandes Fournisseurs"
                ],
                "summary": "Modifier une commande en attente",
                "operationId": "8eecac70a1bc49b8ae2a0b4fa10a953e",
                "responses": {
                    "200": {
                        "description": "Commande mise à jour"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Commandes Fournisseurs"
                ],
                "summary": "Supprimer une commande",
                "operationId": "ab10f768eae05e0ac2d005e8e20a9e60",
                "responses": {
                    "200": {
                        "description": "Commande supprimée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/purchase-orders/{id}/cancel": {
            "post": {
                "tags": [
                    "Commandes Fournisseurs"
                ],
                "summary": "Annuler une commande (Sans mouvement de stock)",
                "operationId": "fa8ff13a2513a56ad4b347f7d177e7ba",
                "responses": {
                    "200": {
                        "description": "Action effectuée avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/purchase-orders/{id}/validate": {
            "post": {
                "tags": [
                    "Commandes Fournisseurs"
                ],
                "summary": "Réceptionner la commande (Entrée en stock)",
                "operationId": "dc705d50c2385d09f4b27f0ab0d59136",
                "responses": {
                    "200": {
                        "description": "Action effectuée avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/purchase-orders/export/pdf": {
            "get": {
                "tags": [
                    "Commandes Fournisseurs"
                ],
                "summary": "Exporter l'historique détaillé en PDF",
                "operationId": "b3e56624c0f300d987346d7312006fc4",
                "responses": {
                    "200": {
                        "description": "Fichier généré avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/purchase-orders/export/excel": {
            "get": {
                "tags": [
                    "Commandes Fournisseurs"
                ],
                "summary": "Exporter l'historique détaillé en Excel",
                "operationId": "63aae7a2fc7f44eece1260569a6a1324",
                "responses": {
                    "200": {
                        "description": "Fichier généré avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Saisissez le jeton (Token) obtenu lors de la connexion pour tester les routes protégées.",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}